requirements.txt
albumentations==2.0.8
opencv-python==4.12.0.88
numpy==2.2.6
pillow==11.3.0
PyYAML==6.0.3
tqdm==4.67.1
matplotlib==3.10.7

torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1

ultralytics==8.3.228
ultralytics-thop==2.0.18
